# docs/api/openapi/knowledge-base-api.yaml
openapi: 3.0.0
info:
  title: 知识库 API
  description: 知识库管理相关接口
  version: 1.0.0
paths:
...
  /api/knowledge-bases/{knowledgeBaseId}/documents:
    get:
      summary: 获取知识库文档列表
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 文档列表获取成功
    post:
      summary: 向知识库添加文档
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                documentId:
                  type: string
      responses:
        '201':
          description: 文档添加成功